+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * kubernetes: add support for managing nodepools
 * kubernetes: add support for managing oidc fields
 * kubernetes: add support for managing servicevm fields
+* kubernetes: add support for managing service user
 
 <!--
 Please add your changelog entry under this comment in the correct category (Security, Fixed, Added, Changed, Deprecated, Removed - in this order).